+++ client/coral-embed-stream/src/containers/LoginContainer.js
