+++ client/coral-embed-stream/src/tabs/stream/components/TagLabel.js
