+++ client/coral-embed-stream/src/tabs/stream/components/__tests__/InfoBox.spec.js
