--- client/talk-plugin-moderation/index.js
@@ -1,1 +1,0 @@
-export { default as ModerationLink } from './ModerationLink';
